Fix codacy issues
Moved copy_but_actually_move into its own impl file
