+++ CMakeLists.txt
@@ -29,6 +29,8 @@ message("${PROJECT_NAME} LIFT_CODE_COVERAGE       = ${LIFT_CODE_COVERAGE}")
 message("${PROJECT_NAME} LIFT_USER_LINK_LIBRARIES = ${LIFT_USER_LINK_LIBRARIES}")
 
 set(LIBLIFTHTTP_SOURCE_FILES
+    inc/lift/impl/copy_util.hpp
+
     inc/lift/client.hpp src/client.cpp
     inc/lift/const.hpp
     inc/lift/escape.hpp src/escape.cpp
